# Improve performance by reduce loadings of Translator