--- a/src/Coduo/PHPHumanizer/Translator/Builder.php
+++ b/src/Coduo/PHPHumanizer/Translator/Builder.php
@@ -16,21 +16,30 @@ use Symfony\Component\Translation\Translator;
 
 final class Builder
 {
+    /**
+     * @var array<string, Translator>
+     */
+    private static array $translators = [];
+
     public static function build(string $locale) : Translator
     {
-        $translator = new Translator($locale);
-        $translator->addLoader('yml', new YamlFileLoader());
+        if (!isset(self::$translators[$locale])) {
+            $translator = new Translator($locale);
+            $translator->addLoader('yml', new YamlFileLoader());
 
-        $iterator = new \FilesystemIterator(__DIR__ . '/../Resources/translations');
-        $filter = new \RegexIterator($iterator, '/[aA-zZ]+\.([a-z]{2}|[a-z]{2}\_[A-Z]{2})\.yml$/');
+            $iterator = new \FilesystemIterator(__DIR__ . '/../Resources/translations');
+            $filter = new \RegexIterator($iterator, '/[aA-zZ]+\.([a-z]{2}|[a-z]{2}\_[A-Z]{2})\.yml$/');
 
-        foreach ($filter as $file) {
-            /* @var $file \SplFileInfo */
-            $resourceName = $file->getBasename('.yml');
-            [$fileDomain, $fileLocale] = \explode('.', $resourceName);
-            $translator->addResource('yml', $file->getPathname(), $fileLocale, $fileDomain);
+            foreach ($filter as $file) {
+                /* @var $file \SplFileInfo */
+                $resourceName = $file->getBasename('.yml');
+                [$fileDomain, $fileLocale] = \explode('.', $resourceName);
+                $translator->addResource('yml', $file->getPathname(), $fileLocale, $fileDomain);
+            }
+
+            self::$translators[$locale] = $translator;
         }
 
-        return $translator;
+        return self::$translators[$locale];
     }
 }