<!-- Bellow section will be used to automatically generate changelog, please do not modify HTML code structure -->
<h2>Change Log</h2>
<div id="change-log">
  <h4>Added</h4>
  <ul id="added">
    <!-- <li>Feature making everything better</li> -->
  </ul> 
  <h4>Fixed</h4>  
  <ul id="fixed">
    <!-- <li>Behavior that was incorrect</li> -->
  </ul>
  <h4>Changed</h4>
  <ul id="changed">
    <li>load Translator only once for given locale</li>
  </ul>  
  <h4>Removed</h4>
  <ul id="removed">
    <!-- <li>Something</li> -->
  </ul>
  <h4>Deprecated</h4>
  <ul id="deprecated">
    <!-- <li>Something is from now deprecated</li> -->
  </ul>  
  <h4>Security</h4> 
  <ul id="security">
    <!-- <li>Something that was security issue, is not an issue anymore</li> -->
  </ul>     
</div>
<hr/>

<h2>Description</h2>

Before:
```
Wall Time     644ms
I/O Wait     3.61ms
CPU Time      640ms
Memory       1.67MB
Network         n/a     n/a     n/a
SQL             n/a     n/a
```

After:
```
Wall Time    22.8ms
I/O Wait       59µs
CPU Time     22.8ms
Memory       1.46MB
Network         n/a     n/a     n/a
SQL             n/a     n/a
```